% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aws_parameters_json_rds.R
\name{aws_parameters_json_minutes}
\alias{aws_parameters_json_minutes}
\title{Create AWS coordinates and parameters JSON file for minutes data.}
\usage{
aws_parameters_json_minutes(aws_dir, nb_col_coords = 7)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}

\item{nb_col_coords}{number of the columns from the coordinates table to take account.}
}
\description{
Create AWS coordinates and parameters JSON file for minutes data.
}
