% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_aws_hourly.R
\name{compute_aws_hourly}
\alias{compute_aws_hourly}
\title{Compute AWS hourly data.}
\usage{
compute_aws_hourly(aws_dir, start_date, end_date)
}
\arguments{
\item{aws_dir}{full path to the folder that will contain the folder AWS_DATA.}

\item{start_date}{the start date of the time series to be computed, in the form "YYYY-mm-dd HH:MM:SS".}

\item{end_date}{the end date of the time series to be computed, in the form "YYYY-mm-dd HH:MM:SS".}
}
\description{
Compute AWS hourly data from minutes time series and write the data into ADT database.
}
