% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_aws_daily.R
\name{compute_one_aws_daily}
\alias{compute_one_aws_daily}
\title{Compute daily data for one AWS.}
\usage{
compute_one_aws_daily(
  aws_dir,
  aws_net,
  aws_id,
  start_date,
  end_date,
  daily_rain_obs_hour = 8,
  first_writing = TRUE
)
}
\arguments{
\item{aws_dir}{full path to the folder that will contain the folder AWS_DATA.}

\item{aws_net}{integer, the network code of the AWS.}

\item{aws_id}{the ID of the AWS.}

\item{start_date}{the start date of the time series to be computed, in the form "YYYY-mm-dd HH:MM:SS".}

\item{end_date}{the end date of the time series to be computed, in the form "YYYY-mm-dd HH:MM:SS".}

\item{daily_rain_obs_hour}{observation hour for daily rainfall data.}

\item{first_writing}{if \code{TRUE} write the data to the database for the first time, \code{FALSE} update existing table.}
}
\description{
Compute daily data for one AWS from hourly time series and write the data into ADT database.
}
