% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_aws_hourly.R
\name{compute_one_aws_hourly}
\alias{compute_one_aws_hourly}
\title{Compute hourly data for one AWS.}
\usage{
compute_one_aws_hourly(aws_dir, aws_net, aws_id, start_date, end_date)
}
\arguments{
\item{aws_dir}{full path to the folder that will contain the folder AWS_DATA.}

\item{aws_net}{integer, the network code of the AWS.}

\item{aws_id}{the ID of the AWS.}

\item{start_date}{the start date of the time series to be computed, in the form "YYYY-mm-dd HH:MM:SS".}

\item{end_date}{the end date of the time series to be computed, in the form "YYYY-mm-dd HH:MM:SS".}
}
\description{
Compute hourly data for one AWS from minutes time series and write the data into ADT database.
}
