% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_aws_coords_tables.R
\name{create_adt_metadata_table}
\alias{create_adt_metadata_table}
\title{Create AWS Metadata tables for all AWS networks.}
\usage{
create_adt_metadata_table(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Create and populate AWS coordinates tables for all AWS networks.
}
