% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_adt_params_table.R
\name{create_adt_params_table}
\alias{create_adt_params_table}
\title{Create ADT parameters table.}
\usage{
create_adt_params_table(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Create ADT parameters table.
}
