% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_adt.R
\name{create_AWS_DATA}
\alias{create_AWS_DATA}
\title{Create ADT AWS data folder.}
\usage{
create_AWS_DATA(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the folder that will contain the folder AWS_DATA.}
}
\description{
Create the folder that will contain ADT data and configuration.
}
