% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadb_timestep_table.R
\name{create_aws_timestep}
\alias{create_aws_timestep}
\title{Create AWS time step information.}
\usage{
create_aws_timestep(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Create AWS time step information.
}
