% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_adt.R
\name{create_db_connection}
\alias{create_db_connection}
\title{Create the authentication to ADT database.}
\usage{
create_db_connection(aws_dir, config_list, file_name)
}
\arguments{
\item{aws_dir}{full path to the folder that will contain the folder AWS_DATA.}

\item{config_list}{a list containing the args of the connection.}

\item{file_name}{the file name of the connection, e.g. 'adt.con'.}
}
\description{
Create the authentication to ADT database.
}
