% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utilities.R
\name{create_db_partition_years}
\alias{create_db_partition_years}
\title{Create MySQL RANGE Partitions.}
\usage{
create_db_partition_years(
  conn,
  table_name,
  col_name,
  partition_years,
  time_thres = "mins"
)
}
\arguments{
\item{conn}{A \code{DBIConnection} object, as produced by \code{\link[DBI]{dbConnect}}.\cr}

\item{table_name}{name of the table.\cr}

\item{col_name}{the name of the column containing the timestamp data.\cr}

\item{partition_years}{a integer vector containing the years to be partitioned.}

\item{time_thres}{the time step of the threshold to be used. Must be one of 'secs', 'mins', 'hours' or 'days'.}
}
\description{
Create MySQL range partitions by year on an existing table.
}
