% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utilities.R
\name{create_db_table}
\alias{create_db_table}
\title{Create MySQL table.}
\usage{
create_db_table(conn, table_name, ...)
}
\arguments{
\item{conn}{A \code{DBIConnection} object, as produced by \code{\link[DBI]{dbConnect}}.\cr}

\item{table_name}{name of the table.\cr}

\item{...}{definition of the table columns. For more details of the create definition arguments, go to \href{https://dev.mysql.com/doc/refman/8.0/en/create-table.html}{MySQL CREATE TABLE Statement}}
}
\description{
Create MySQL table with a given name in ADT database.
}
