% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_minfrac_tables.R
\name{create_minfrac_daily}
\alias{create_minfrac_daily}
\title{Create minimum fraction table to compute daily time series.}
\usage{
create_minfrac_daily(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Create minimum fraction table to compute daily time series.
}
