% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network_infos.R
\name{create_network_info}
\alias{create_network_info}
\title{Create AWS network information.}
\usage{
create_network_info(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Create AWS network information, write to a JSON file and to ADT database.
}
