% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aws_parameters_json_rds.R
\name{get_aws_parameters_rds}
\alias{get_aws_parameters_rds}
\title{Create AWS coordinates and parameters RDS file.}
\usage{
get_aws_parameters_rds(aws_dir, data_timestep, nb_col_coords = 7)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}

\item{data_timestep}{the time step of the data. Available options: "minutes", "hourly" or "daily".}

\item{nb_col_coords}{number of the columns from the coordinates table to take account.}
}
\description{
Create AWS coordinates and parameters RDS file.
}
