% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadb_range.R
\name{get_datadb_range}
\alias{get_datadb_range}
\title{Date range of the available observation.}
\usage{
get_datadb_range(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Get the date rage of the available observation for each AWS.
}
