% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadb_timestep.R
\name{get_datadb_timestep}
\alias{get_datadb_timestep}
\title{Time step of each AWS.}
\usage{
get_datadb_timestep(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Get the time step of each AWS.
}
