% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_aws_status.R
\name{initializeAWSStatus}
\alias{initializeAWSStatus}
\title{Initialize AWS status.}
\usage{
initializeAWSStatus(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Create an initial AWS status rds file.
}
