% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_aws_minutes.R
\name{populate_one_aws_minutes}
\alias{populate_one_aws_minutes}
\title{Write minutes data into the database for one AWS.}
\usage{
populate_one_aws_minutes(
  aws_dir,
  aws_net,
  aws_id,
  start_date,
  end_date,
  data_dir = NULL
)
}
\arguments{
\item{aws_dir}{full path to the folder that will contain the folder AWS_DATA.}

\item{aws_net}{integer, the network code of the AWS.}

\item{aws_id}{the ID of the AWS.}

\item{start_date}{the start date of the time series to be populated, in the form "YYYY-mm-dd HH:MM:SS".}

\item{end_date}{the end date of the time series to be populated, in the form "YYYY-mm-dd HH:MM:SS".}

\item{data_dir}{if the data is stored in folder outside the DATA folder in AWS_DATA,\cr
full path to the folder that will contain the folder of the AWS network.\cr
The name of the AWS network folder must be the same as the names in the column \code{name_dir} of the file \code{CSV/adt_network_table.csv}.}
}
\description{
Populate minutes data into ADT database for one AWS.
}
