% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadb_range.R
\name{reinit_table_crds_aws_aggr_ts}
\alias{reinit_table_crds_aws_aggr_ts}
\title{Populate the AWS time range for the metadata and 'aws_aggr_ts' tables .}
\usage{
reinit_table_crds_aws_aggr_ts(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Populate the start and end date of each AWS network metadata tables and aws_aggr_ts table.
}
