% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_adt_colorkey_tables.R
\name{update_adt_colorkey_table}
\alias{update_adt_colorkey_table}
\title{Update ADT colorkey table.}
\usage{
update_adt_colorkey_table(aws_dir, csv_file, table_name)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}

\item{csv_file}{the csv file containing the colorkey table located under the folder CSV.}

\item{table_name}{name of the table.}
}
\description{
Update ADT colorkey table.
}
