% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_limit_check_table.R
\name{update_adt_limit_check_table}
\alias{update_adt_limit_check_table}
\title{Update limit check table.}
\usage{
update_adt_limit_check_table(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Update the table containing the minimum and maximum limits to be used on minutes QC for each AWS and each parameter.
}
