% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_adt_params_table.R
\name{update_adt_params_table}
\alias{update_adt_params_table}
\title{Update ADT parameters table.}
\usage{
update_adt_params_table(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Update ADT parameters table.
}
