% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_adt_params_table.R
\name{update_adt_stats_table}
\alias{update_adt_stats_table}
\title{Update ADT statistics table.}
\usage{
update_adt_stats_table(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Update ADT statistics table.
}
