% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_aws_daily.R
\name{update_aws_daily}
\alias{update_aws_daily}
\title{Update AWS daily data.}
\usage{
update_aws_daily(
  aws_dir,
  daily_rain_obs_hour = 8,
  start_time = "2015-01-01 00:00:00"
)
}
\arguments{
\item{aws_dir}{full path to the folder that will contain the folder AWS_DATA.}

\item{daily_rain_obs_hour}{observation hour for daily rainfall data.}

\item{start_time}{for new station, the start time the computation will be started.}
}
\description{
Compute AWS daily data from hourly time series and write the data into ADT database.
}
