% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_aws_hourly.R
\name{update_aws_hourly}
\alias{update_aws_hourly}
\title{Update AWS hourly data.}
\usage{
update_aws_hourly(
  aws_dir,
  start_time = "2015-01-01 00:00:00",
  first_writing = FALSE
)
}
\arguments{
\item{aws_dir}{full path to the folder that will contain the folder AWS_DATA.}

\item{start_time}{for new station, the start time the computation will be started.}

\item{first_writing}{if \code{TRUE} write the data to the database for the first time, \code{FALSE} update existing table.}
}
\description{
Compute AWS hourly data from minutes time series and write the data into ADT database.
}
