% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_aws_minutes.R
\name{update_aws_minutes}
\alias{update_aws_minutes}
\title{Update AWS minutes data.}
\usage{
update_aws_minutes(
  aws_dir,
  start_time = "2015-01-01 00:00:00",
  first_writing = FALSE
)
}
\arguments{
\item{aws_dir}{full path to the folder that will contain the folder AWS_DATA.}

\item{start_time}{for new station, the start time of the data.}

\item{first_writing}{if \code{TRUE} write the data to the database for the first time, \code{FALSE} update existing table.}
}
\description{
Updates AWS minutes data from minutes time series and write the data into ADT database.
}
