% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadb_timestep_table.R
\name{update_aws_timestep}
\alias{update_aws_timestep}
\title{Update AWS time step information.}
\usage{
update_aws_timestep(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Update AWS time step information.
}
