% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_minfrac_tables.R
\name{update_minfrac_hourly}
\alias{update_minfrac_hourly}
\title{Create minimum fraction table to compute hourly time series.}
\usage{
update_minfrac_hourly(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Create minimum fraction table to compute hourly time series.
}
