% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_aws_coords_tables.R
\name{update_net_metadata_table}
\alias{update_net_metadata_table}
\title{Update AWS Metadata table for one AWS network.}
\usage{
update_net_metadata_table(aws_dir, table_name)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}

\item{table_name}{name of the table, must be the same name as defined in the table \code{adt_network}.}
}
\description{
Update AWS coordinates table for one AWS network.
}
