% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network_infos.R
\name{update_network_info}
\alias{update_network_info}
\title{Update AWS network information.}
\usage{
update_network_info(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Update AWS network information.
}
