% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_aws_status.R
\name{updateAWSStatus}
\alias{updateAWSStatus}
\title{Update AWS data status.}
\usage{
updateAWSStatus(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder.}
}
\description{
Update AWS data status every hour.
}
